Fix to compile with C++

+++ examples/main.c
@@ -720,7 +720,7 @@ int main(int argc, char *argv[]) {
     /* Get the total number of particles across all nodes. */
     long long N_total[3] = {0, 0, 0};
 #if defined(WITH_MPI)
-    long long N_long[3] = {Ngas, Ngpart, Nspart};
+    long long N_long[3] = {(long long)Ngas, (long long)Ngpart, (long long)Nspart};
     MPI_Allreduce(&N_long, &N_total, 3, MPI_LONG_LONG_INT, MPI_SUM,
                   MPI_COMM_WORLD);
 #else
